requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.2.0
celery==5.3.6
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==5.0.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
idna==3.6
kombu==5.3.5
librabbitmq==2.0.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
python-dateutil==2.8.2
pytz==2023.4
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.2.0
vine==5.1.0
wcwidth==0.2.13
